chore: extract out PostgrestBuilder
