+++ package.json
@@ -45,7 +45,7 @@
     "semantic-release-plugin-update-version-in-files": "^1.1.0",
     "ts-jest": "^26.3.0",
     "typedoc": "^0.22.15",
-    "typescript": "^4.0.2",
+    "typescript": "^4.6.4",
     "wait-for-localhost-cli": "^2.0.0"
   }
 }
