+++ src/index.ts
@@ -1,21 +1,11 @@
-import PostgrestClient from './PostgrestClient'
-import PostgrestFilterBuilder from './lib/PostgrestFilterBuilder'
-import PostgrestQueryBuilder from './lib/PostgrestQueryBuilder'
-import {
-  PostgrestBuilder,
+export { default as PostgrestClient } from './PostgrestClient'
+export { default as PostgrestQueryBuilder } from './lib/PostgrestQueryBuilder'
+export { default as PostgrestFilterBuilder } from './lib/PostgrestFilterBuilder'
+export { default as PostgrestTransformBuilder } from './lib/PostgrestTransformBuilder'
+export { default as PostgrestBuilder } from './lib/PostgrestBuilder'
+export {
   PostgrestResponse,
   PostgrestSingleResponse,
   PostgrestMaybeSingleResponse,
   PostgrestError,
 } from './lib/types'
-
-export {
-  PostgrestClient,
-  PostgrestBuilder,
-  PostgrestQueryBuilder,
-  PostgrestFilterBuilder,
-  PostgrestResponse,
-  PostgrestSingleResponse,
-  PostgrestMaybeSingleResponse,
-  PostgrestError,
-}
